requirements.txt
kaggle
torch
torchvision
gradio
Pillow
tqdm
numpy
tensorboard 
boto3